requirements.txt
google-adk
google-generativeai
google-cloud-logging

fastapi
uvicorn[standard]
python-multipart
aiofiles

httpx
pydantic
python-dotenv